refactor: Split Tier 2 medium-impact files into focused modules
Split 3 medium-sized components (580-714 lines each) into smaller,
focused modules.

- PreferencesPage.tsx (714→40 lines): Extract ProfileSettings,
  PasswordSettings, FontSettings, AppearanceSettings
- CollapsibleSidebar.tsx (714→410 lines): Extract SidebarRail,
  useResizeHandle, useSectionToggle
- AdminSettingsPage.tsx (583→21 lines): Extract
  AdminSystemSettings, AdminUserManagement
